Rewrote the chapter navigation
The chapter navigation was slapped together in the easiest way possible. Just copy and paste each outcome with toggling on and off 'block' and 'none'. I didn't want to think about the function that was more efficient at the time, but I was able to get some free time and rewrite it. It now has two functions instead of 16 for navigating the 16 chapters. This saved around 450 lines of code, woohoo!